requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
APScheduler==3.11.2
blinker==1.9.0
cachetools==7.0.3
category_encoders==2.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cloudpickle==3.1.2
colorlog==6.10.1
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
databricks-sdk==0.96.0
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
dynaconf==3.2.12
evidently==0.7.20
Faker==40.8.0
fastapi==0.135.1
filelock==3.25.0
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.61.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.0
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huey==2.6.0
idna==3.11
importlib_metadata==8.7.1
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lightgbm==4.6.0
litestar==2.21.1
litestar-htmx==0.5.0
loguru==0.7.3
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mlflow==3.10.1
mlflow-skinny==3.10.1
mlflow-tracing==3.10.1
msgspec==0.20.0
multidict==6.7.1
multipart==1.3.1
mypy_extensions==1.1.0
nltk==3.9.3
numpy==2.4.2
opentelemetry-api==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
optuna==4.7.0
packaging==26.0
pandas==2.3.3
pandera==0.29.0
patsy==1.0.2
pillow==12.1.1
plotly==5.24.1
polyfactory==3.3.0
prettytable==3.17.0
protobuf==6.33.5
psycopg2-binary==2.9.11
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rich-click==1.9.7
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
shellingham==1.5.4
six==1.17.0
skops==0.13.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlparse==0.5.5
starlette==0.52.1
statsmodels==0.14.6
tenacity==9.1.4
threadpoolctl==3.6.0
tqdm==4.67.3
typeguard==4.5.1
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.3
uuid6==2025.0.1
uvicorn==0.41.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
Werkzeug==3.1.6
xgboost==3.2.0
zipp==3.23.0
